Update README.md
1 Introduction

Artificial Neural networks (ANN) are inspired by the human brain structure. They are composed of several interconnected units, called neurons, capable of a single input-to-output trans-formation. There are usually several connections, and their properties are summarized in parameters called weights. In a Neural Network, neurons are organized in layers cascaded and connected to each other by patterns.

2 Theory of Neural Networks

A Neural Network is composed of a set of neurons. A neuron is a single unit that takes an input, makes a decision, and returns an output. It can receive inputs either from the outside world or from another neuron in the previous layer. Even the output of a neuron can be directed to the outside world or to a neuron in the following layer.

Each neuron is associated with a weight (w), that changes dur-ing the learning process accordingly to predefined rules, de-fined by the weight update procedure: